Add multi-timestamp to rdump Remove multi-timestamp, add multi-timestamp function to rdump and adjust tests
